refactor(queue): 拆分MessageQueuePanel为统计+列表+迷你图+刷新策略
